form fix

+++ controllers/restaurants.js
@@ -68,9 +68,14 @@ module.exports.editRestaurant = async (req, res) => {
 }
 
 module.exports.delete = async (req, res) => {
-    await Restaurant.findByIdAndDelete(req.params.id);
-    alert("確定刪除？")
-    req.flash('fail', '成功刪除餐廳');
-    res.redirect('/restaurants');
+    const yes = confirm("確定刪除？");
+    if (yes) {
+        await Restaurant.findByIdAndDelete(req.params.id);
+        req.flash('fail', '成功刪除餐廳');
+        res.redirect('/restaurants');
+    } else {
+        res.redirect(`/restaurants/${req.params.id}`);
+    }
+
 }
